requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.35.10
botocore==1.35.10
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-tailwind==1.0.3
dj-database-url==2.3.0
Django==4.2
django-cors-headers==4.4.0
django-crispy-forms==1.14.0
django-filter==24.3
django-mathfilters==1.0.0
django-storages==1.13.1
django-tailwind==3.6.0
django-utils-six==2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dotenv==0.9.9
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
packaging==24.2
pillow==10.3.0
psycopg==3.2.5
psycopg-binary==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
whitenoise==6.7.0
